# internal/domain/elevator/operation_state.go
package elevator

type OperationState string

const (
	OperationStateRunning     OperationState = "running"
	OperationStateStopped     OperationState = "stopped"
	OperationStateMaintenance OperationState = "maintenance"
)